refactor: move user related functions from _helpers.js to utils/users.js
