+++ controllers/followshipController.js
@@ -1,7 +1,6 @@
 const db = require('../models')
 const Followship = db.Followship
 const User = db.User
-const helpers = require('../_helpers')
 
 const followshipController = {
   followUser: async (req, res, next) => {
@@ -16,21 +15,17 @@ const followshipController = {
       })
 
       if (!followingUser) {
-        return res
-          .status(200)
-          .json({
-            status: 'error',
-            message: 'this user doesn\'t exist'
-          })
+        return res.status(200).json({
+          status: 'error',
+          message: "this user doesn't exist"
+        })
       }
 
       if (Number(followingId) === followerId) {
-        return res
-          .status(403)
-          .json({
-            status: 'error',
-            message: 'You cannot follow yourself.'
-          })
+        return res.status(403).json({
+          status: 'error',
+          message: 'You cannot follow yourself.'
+        })
       }
 
       const followship = await Followship.findOne({
@@ -41,12 +36,10 @@ const followshipController = {
       })
 
       if (followship) {
-        return res
-          .status(409)
-          .json({
-            status: 'error',
-            message: `already followed @${followingUser.account}`
-          })
+        return res.status(409).json({
+          status: 'error',
+          message: `already followed @${followingUser.account}`
+        })
       }
 
       await Followship.create({
@@ -54,16 +47,11 @@ const followshipController = {
         followingId
       })
 
-      return res
-        .status(200)
-        .json({
-          status: 'success',
-          message: `followed @${followingUser.account}`,
-          followingUser
-        })
-
-
-
+      return res.status(200).json({
+        status: 'success',
+        message: `followed @${followingUser.account}`,
+        followingUser
+      })
     } catch (error) {
       next(error)
     }
@@ -81,21 +69,17 @@ const followshipController = {
       })
 
       if (!unfollowedUser) {
-        return res
-          .status(200)
-          .json({
-            status: 'error',
-            message: 'cannot unfollow an user that doesn\'t exist'
-          })
+        return res.status(200).json({
+          status: 'error',
+          message: "cannot unfollow an user that doesn't exist"
+        })
       }
 
       if (Number(followingId) === followerId) {
-        return res
-          .status(403)
-          .json({
-            status: 'error',
-            message: 'You cannot unfollow yourself.'
-          })
+        return res.status(403).json({
+          status: 'error',
+          message: 'You cannot unfollow yourself.'
+        })
       }
 
       const followship = await Followship.findOne({
@@ -106,22 +90,18 @@ const followshipController = {
       })
 
       if (!followship) {
-        return res
-          .status(200)
-          .json({
-            status: 'error',
-            message: 'cannot unfollow since you haven\'t followed this user before'
-          })
+        return res.status(200).json({
+          status: 'error',
+          message: "cannot unfollow since you haven't followed this user before"
+        })
       }
 
       await followship.destroy()
 
-      return res
-        .status(200)
-        .json({
-          status: 'success',
-          message: `Unfollowed ${unfollowedUser.account}`
-        })
+      return res.status(200).json({
+        status: 'success',
+        message: `Unfollowed ${unfollowedUser.account}`
+      })
     } catch (error) {
       next(error)
     }
